plumbing: format/commitgraph, Convert v2 into commitgraph Some previous changes to the commitgraph package required a breaking the API, which resulted in the /v2 code.
Refer to #869 for additional context.
Relates to #910.
