+++ plumbing/format/commitgraph/chain.go
@@ -1,4 +1,4 @@
-package v2
+package commitgraph
 
 import (
 	"bufio"
