+++ plumbing/format/commitgraph/chain_test.go
@@ -1,11 +1,11 @@
-package v2_test
+package commitgraph_test
 
 import (
 	"bytes"
 	"crypto"
 	"strings"
 
-	commitgraph "github.com/go-git/go-git/v6/plumbing/format/commitgraph/v2"
+	"github.com/go-git/go-git/v6/plumbing/format/commitgraph"
 	"github.com/go-git/go-git/v6/plumbing/hash"
 )
